Add function "Pause" and dynamic level of complexity Add tests

+++ src/project/costruction/code/Gameplay.java
@@ -7,7 +7,7 @@ import java.awt.event.KeyEvent;
 import java.util.Arrays;
 import java.util.Random;
 
-public class Gameplay extends JFrame{
+public class Gameplay extends JFrame {
 
     final String TITLE_OF_PROGRAM = "Tetris"; //название окна
     static final int BLOCK_SIZE = 25; //размер одного блока
@@ -21,7 +21,8 @@ public class Gameplay extends JFrame{
     static final int UP = 38; //вверх
     static final int RIGHT = 39; // вправо
     static final int DOWN = 40; // вниз
-    static final int SHOW_DELAY = 350; // задержка хода
+    static int SHOW_DELAY = 500; // задержка хода
+    int currentScore = 0;
     static final int[][][] SHAPES = { //массив всех фигур
             {{0, 0, 0, 0}, {1, 1, 1, 1}, {0, 0, 0, 0}, {0, 0, 0, 0}, {4, 0x00f0f0}}, // I
             {{0, 0, 0, 0}, {0, 1, 1, 0}, {0, 1, 1, 0}, {0, 0, 0, 0}, {4, 0xf0f000}}, // O
@@ -36,6 +37,7 @@ public class Gameplay extends JFrame{
     static int[][] mine = new int[FIELD_HEIGHT + 1][FIELD_WIDTH]; // поле длля игры
     JFrame frame; //наши окна
     Canvas canvas = new Canvas();
+    boolean isPaused = false;
     static Figure figure = new Figure();
     static boolean gameOver = false;
     static final int[][] GAME_OVER_MSG = { //код для конца игры
@@ -51,6 +53,7 @@ public class Gameplay extends JFrame{
             {1, 0, 0, 1, 0, 1, 0, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0},
             {1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0},
             {0, 1, 1, 0, 0, 1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0}};
+
     public Gameplay() {
         setTitle(TITLE_OF_PROGRAM); //задаем название окна
         setDefaultCloseOperation(EXIT_ON_CLOSE); // задаем для закрытия
@@ -59,11 +62,13 @@ public class Gameplay extends JFrame{
         canvas.setBackground(Color.black); // задаем цвет задника
         addKeyListener(new KeyAdapter() { //обработчик нажатий клавиш
             public void keyPressed(KeyEvent e) {
-                if (!gameOver) {
+                if (!gameOver && !isPaused) {
                     if (e.getKeyCode() == DOWN) figure.drop();
                     if (e.getKeyCode() == UP) figure.rotate();
                     if (e.getKeyCode() == LEFT || e.getKeyCode() == RIGHT) figure.move(e.getKeyCode());
                 }
+                if (e.getKeyCode() == 32)
+                    isPaused = !isPaused;
                 canvas.repaint();
             }
         });
@@ -76,15 +81,23 @@ public class Gameplay extends JFrame{
         while (!gameOver) {
             try {
                 Thread.sleep(SHOW_DELAY);
-            } catch (Exception e) { e.printStackTrace(); }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            if (currentScore >= 1000) {
+                currentScore -= 1000;
+                SHOW_DELAY /= 1.2;
+            }
             canvas.repaint();
             checkFilling();
-            if (figure.isTouchGround()) {
-                figure.leaveOnTheGround();
-                figure = new Figure();
-                gameOver = figure.isCrossGround(); // Is there space for a new figure?
-            } else
-                figure.stepDown();
+            if (!isPaused) {
+                if (figure.isTouchGround()) {
+                    figure.leaveOnTheGround();
+                    figure = new Figure();
+                    gameOver = figure.isCrossGround(); // Is there space for a new figure?
+                } else
+                    figure.stepDown();
+            }
         }
     }
 
@@ -97,12 +110,13 @@ public class Gameplay extends JFrame{
                 filled *= Integer.signum(mine[row][col]);
             if (filled > 0) {
                 countFillRows++;
-                for (int i = row; i > 0; i--) System.arraycopy(mine[i-1], 0, mine[i], 0, FIELD_WIDTH);
+                for (int i = row; i > 0; i--) System.arraycopy(mine[i - 1], 0, mine[i], 0, FIELD_WIDTH);
             } else
                 row--;
         }
         if (countFillRows > 0) {
             gameScore += SCORES[countFillRows - 1];
+            currentScore += SCORES[countFillRows - 1];
             setTitle(TITLE_OF_PROGRAM + " : " + gameScore);
         }
     }
